2.0.0

+++ ViewModel/ReceiptRequestModel.cs
@@ -1,8 +1,5 @@
 ﻿using SPF_Receipt.DataModel;
-using System;
 using System.Collections.Generic;
-using System.Linq;
-using System.Threading.Tasks;
 
 namespace SPF_Receipt.ViewModel
 {
@@ -10,7 +7,7 @@ namespace SPF_Receipt.ViewModel
     {
         public ReportHeaderModel ReceiptHeader { get; set; }
         public Customer Customer { get; set; }
-        public IEnumerable<ReportDataModel<double, int>> ProductList { get; set; }
+        public IEnumerable<ReportDataModel<double, decimal>> ProductList { get; set; }
         public ReportFooterModel<decimal> ReceiptSummary { get; set; }
     }
